[Defend Workflows][E2E]Endpoint e2e response console (#155605)
Depends on https://github.com/elastic/kibana/pull/155519

E2E coverage of `isolate`, `processes`, `kill-process` and
`suspend-process` commands on mocked endpoint.

E2E coverage of the above but on real endpoint is
[here](https://github.com/elastic/kibana/pull/155519).

Because these tests are run against mocked data I've decided not to mock
`kill-process` and `suspend-process` outcome (whether process is
actually killed/suspended) because it would mean testing mocks
themselves. What is tested is the outcome the user sees ('Action
completed').

---------

(cherry picked from commit fd5309f6a02bce641c4baf79500acfe797e294f7)
