+++ x-pack/plugins/security_solution/public/management/cypress/tasks/response_console.ts
@@ -7,6 +7,7 @@
 
 import { closeAllToasts } from './close_all_toasts';
 import { APP_ENDPOINTS_PATH } from '../../../../common/constants';
+import Chainable = Cypress.Chainable;
 
 export const waitForEndpointListPageToBeLoaded = (endpointHostname: string): void => {
   cy.visit(APP_ENDPOINTS_PATH);
@@ -55,4 +56,17 @@ export const performCommandInputChecks = (command: string) => {
   clearConsoleCommandInput();
   selectCommandFromHelpMenu(command);
   checkInputForCommandPresence(command);
+};
+
+export const checkReturnedProcessesTable = (): Chainable<JQuery<HTMLTableRowElement>> => {
+  ['USER', 'PID', 'ENTITY ID', 'COMMAND'].forEach((header) => {
+    cy.contains(header);
+  });
+
+  return cy
+    .get('tbody')
+    .find('tr')
+    .then((rows) => {
+      expect(rows.length).to.be.greaterThan(0);
+    });
 };
